Remove old storage quota interface
